organizes error files into one folder
