+++ src/app.js
@@ -19,7 +19,7 @@ const {
 } = config;
 
 const redirect_uri = `${host}:${port}/oauth/redirect`;
-const errorMap = require('./errorMap');
+const errorMap = require('./customError/errorMap');
 const parser = require('./parser/index');
 const createAction = require('./actions/index');
 const reduce = require('./reducers/index');
@@ -83,10 +83,7 @@ app.get('/oauth/authorize', (req, res) => {
 });
 
 app.get('/oauth/redirect', async(req, res) => {
-  const {
-    code,
-    state
-  } = req.query;
+  const { code, state } = req.query;
   try {
     await getSecretsAndSave(code);
     res.send('Thank you, you have successfully authenticated your team!');
@@ -114,40 +111,25 @@ app.post('/api/command', (req, res) => {
 
 app.listen(port, () => console.log(`listening on port ${port}!`));
 
-async function respond({
-  team_id,
-  user_id,
-  text,
-  channel_id
-}) {
+async function respond({ team_id, user_id, text, channel_id }) {
   const secret = await getSecret({
     team_id
   });
+
   const token = secret.bot.bot_access_token;
   const command = parser(text);
+
   if (command.type === 'STATUS') {
     const statusMap = await status(team_id);
     await sendMessage(channel_id, token, JSON.stringify(statusMap));
   }
   else {
-    await executeCommand({
-      team_id,
-      user_id,
-      command,
-      token
-    });
+    await executeCommand({ team_id, user_id, command, token });
   }
 }
 
-async function executeCommand({
-  team_id,
-  user_id,
-  command,
-  token
-}) {
-  const events = await getEventsFor({
-    team_id
-  });
+async function executeCommand({ team_id, user_id, command, token }) {
+  const events = await getEventsFor({ team_id });
   const eventIds = new Set(events.map(e => e.event_id));
   const usersInfo = await getUsersInfo(token);
   const action = createAction(command, usersInfo, user_id, eventIds, 'UTC');
